Update: Combined into single multion router + added redis db functions
