--- plugins/example/_multion/__init__.py
@@ -1,2 +1,0 @@
-from .router import router
-from .demo_router import demo_router
